Separating info and metrics in zedagent
