+++ pkg/pillar/cmd/zedagent/handlevolume.go
@@ -11,9 +11,7 @@ import (
 	"fmt"
 
 	zconfig "github.com/lf-edge/eve/api/go/config"
-	"github.com/lf-edge/eve/api/go/metrics"
 	"github.com/lf-edge/eve/pkg/pillar/types"
-	"github.com/lf-edge/eve/pkg/pillar/utils"
 	"github.com/satori/go.uuid"
 )
 
@@ -132,38 +130,4 @@ func handleVolumeStatusDelete(ctxArg interface{},
 	uuidStr := status.VolumeID.String()
 	PublishVolumeToZedCloud(ctx, uuidStr, nil, ctx.iteration)
 	ctx.iteration++
-}
-
-func createVolumeInstanceMetrics(ctx *getconfigContext, reportMetrics *metrics.ZMetricMsg) {
-	log.Debugf("Volume instance metrics started")
-	sub := ctx.subVolumeStatus
-	volumelist := sub.GetAll()
-	if volumelist == nil || len(volumelist) == 0 {
-		return
-	}
-	for _, volume := range volumelist {
-		volumeStatus := volume.(types.VolumeStatus)
-		volumeMetric := new(metrics.ZMetricVolume)
-		volumeMetric.Uuid = volumeStatus.VolumeID.String()
-		volumeMetric.DisplayName = volumeStatus.DisplayName
-		if volumeStatus.FileLocation == "" {
-			log.Infof("FileLocation is empty for %s", volumeStatus.Key())
-		} else {
-			getVolumeResourcesMetrics(volumeStatus.FileLocation, volumeMetric)
-		}
-		reportMetrics.Vm = append(reportMetrics.Vm, volumeMetric)
-	}
-	log.Debugf("Volume instance metrics done: %v", reportMetrics.Vm)
-}
-
-func getVolumeResourcesMetrics(name string, volumeMetric *metrics.ZMetricVolume) error {
-
-	actualSize, maxSize, err := utils.GetVolumeSize(name)
-	if err != nil {
-		return err
-	}
-	volumeMetric.UsedBytes = actualSize
-	volumeMetric.TotalBytes = maxSize
-	volumeMetric.FreeBytes = maxSize - actualSize
-	return nil
 }
